Refactor and migrate webhook paths forGitHub
